get private releases working
- clone `glass` with my own credentials
- use catbox.moe for non-public releases

takeaways for my future self:
- gist repositories have a size limit and tend to time out/bail when close to one -- they aren't designed to store binary data after all
- https://github.com/shimataro/ssh-key-action is the most sane and feature-complete SSH setup action
- secret sharing between jobs requires an external service: https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#example-masking-and-passing-a-secret-between-jobs-or-workflows
- https://put.re is good, yet locked down: https://gitlab.com/putre/frontend/-/issues/16
- `GH_TOKEN=... gh api -X PATCH gists/$hash -F "files[$filename][content]=@$filename"` might be the easiest way to update a gist

honestly, fed up with almost every part of the pipeline not working EVER SO SLIGHTLY AS IT SHOULD from the get-go.
even worse, sometimes it takes an immense amount of time and hoops to jump through to get it fixed, OR to come with terms that you wasted your time, as simple as that.

/rant